RISC-V: support for vector register accesses via ptrace() in RISC-V Linux native
Add support for GDB to read from/write to vector registers in a native riscv
linux environment. This uses the NT_RISCV_VECTOR ptrace flag introduced in
Linux 6.5. The values that can be collected by this API are the vector CSRS
vstart, vl, vtype, vcsr, vlenb along with all of the 32 vector registers which
are stored in the pointer datap.

Different from other architectures, riscv vector units can have a variable size
of between 32 bits and 8124 bits for each of the 32 vector registers. All
vector registers on a single system must have the same size. The size of the
vector registers supported by the hardware is exposed by the vlenb csr. As a
consequence, the size of datap will be different depending on the size of the
hardware's vector registers, which GDB is able to determine by using the vlenb
csr.

Co-Authored-By: Greg Savin <greg.savin@sifive.com>
Co-Authored-By: GSoham Gargote <sgargote@whileone.in>

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>